Restructure files from /src/api to /src/lib...
