+++ src/lib/types/externals.ts
